changed prior_time data type from float32 to float64, writing it in float32 can causes about 30 sec to 1 min difference in the times compared to GEOS5 times

+++ write_tccon_netcdf/write_netcdf.py
@@ -1372,7 +1372,7 @@ def main():
         }
         nc_data['time'].setncatts(att_dict)
 
-        nc_data.createVariable('prior_time',np.float32,('prior_time'))
+        nc_data.createVariable('prior_time',np.float64,('prior_time'))
         att_dict = {
             "standard_name": "prior_time",
             "long_name": "prior time",
